:wrench: handle noop `RunAtomically` when mocked
Mocked ops do not return a QueryExecutor, because they are intercepting
the operations at a logical level rather than at the level of statements
and values.

This means that when a gocass.Noop is created as a non-mocked Noop, it
forms a unmocked collection of mocked operations, and continues to use
a standard `RunAtomically` execution, which tries to find a query
executor and passes through the collective statements. A mocked atomic
run just runs the normal run function...

If we allow the creation of Noops to also be mocked, then we can create
a mock noop, so that we have a mocked collection of mocked operations,
which causes delegation of running to occur more-or-less correctly.

One major problem with this approach is that the `gocassa.Noop` is a global
function, so setting this factory globally clobbers the function, and would
mean that if you had a non-mocked keyspace all of its RunAtomically functions
would suddenly become non-atomic. The assumption here is that if you're
mocking a keyspace then you're not doing it piecemeal, i.e. there are not also
concrete keyspaces. This is just an assumption, not a guarantee though--an
alternative approach to this PR would be to make mocking the noop function opt-in
and make the user explicitly choose to do this.