:wrench: handle mock table reads without relations
Presently we attempt to validate the primary key is set in the
query relation--presumably to avoid issuing a query against a
composite key with only a portion of the pk fields. However, it
is overly cautious in the case of there being no relations at
all (which we do rarely to effectively scan a table) and reports
a PK error. Instead when no relations are specified we should
just read all the rows of the mocked table instead.